Remove explicit g++ usage from Makefile
The Makefile for AGM explicitly used "g++" for the C++ compiler.  This
is undesirable, as it makes it more difficult to build on systems
whose native compiler isn't GCC (FreeBSD and Mac OS X use Clang).

This commit replaces the explicit use of "g++" with "$(CXX)", Make's
normal variable that should have the default C++ compiler named in
it.  This can also be used to override which C++ compiler is being
used, via "make CXX=<something else>".