Change CPP variable to CXX, cleanup
The Makefile for SORAL used "$(CPP)" for the C++ compiler and
"$(CPPFLAGS)" for the flags to pass to the compiler.  This is
nonstandard usage in Makefiles, where the CPP variable is used for the
C preprocessor, and CPPFLAGS for the preprocessor flags.  CXX is the
usual variable for Makefiles to refer to the C++ compiler, and
CXXFLAGS its associated flags variable.

This commit brings the Makefile in line with more standard usage.